Move Restriction and Image to top level of model module.
