+++ src/model/mod.rs
@@ -7,7 +7,6 @@ pub mod context;
 pub mod cud_result;
 pub mod device;
 pub mod enums;
-pub mod image;
 pub mod offset;
 pub mod page;
 pub mod playing;
@@ -19,6 +18,22 @@ pub mod track;
 pub mod user;
 
 use serde::{Deserialize, Serialize};
+use RestrictionReason;
+
+/// [image object](https://developer.spotify.com/documentation/web-api/reference/object-model/#image-object)
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
+pub struct Image {
+    pub height: Option<u32>,
+    pub url: String,
+    pub width: Option<u32>,
+}
+
+/// [track restriction object](https://developer.spotify.com/documentation/web-api/reference/object-model/#track-restriction-object)
+/// [album restriction object](https://developer.spotify.com/documentation/web-api/reference/object-model/#album-restriction-object)
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
+pub struct Restriction {
+    pub reason: RestrictionReason,
+}
 
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(untagged)]
@@ -29,6 +44,6 @@ pub enum PlayingItem {
 
 pub use {
     album::*, artist::*, audio::*, category::*, context::*, cud_result::*, device::*, enums::*,
-    image::*, offset::*, page::*, playing::*, playlist::*, recommend::*, search::*, show::*,
-    track::*, user::*,
+    offset::*, page::*, playing::*, playlist::*, recommend::*, search::*, show::*, track::*,
+    user::*,
 };
